Unify the Windows & UNIX spawnSync code paths using `shell: true`
Node.js 4.8.0 & newer include the `shell` config option which fires a command
in a real shell. This makes it no longer necessary to have a separate code path
for Windows & UNIX.

Relying on it exclusively would bump the required Node.js version
to at least 4.8.0, though, so the old code paths are still there
in the code base. We're testing on Node.js 4.0 in addition to latest
4, 6 & 7 to verify this use case is not broken.